Refactor member access
